NMS-14699 fix cyclic reference, rename packages
