+++ client/cache/pom.xml
@@ -8,12 +8,10 @@
         <version>0.1.0-SNAPSHOT</version>
     </parent>
     <modelVersion>4.0.0</modelVersion>
-
-    <groupId>org.opennms.plugins.velocloud</groupId>
-    <artifactId>cache</artifactId>
-    <name>OpenNMS :: Plugins :: Velocloud Intergration :: Cache</name>
-
+    <artifactId>velocloud-client-cache</artifactId>
+    <name>OpenNMS :: Plugins :: Velocloud Intergration :: Client :: Cache</name>
     <packaging>bundle</packaging>
+
     <build>
         <plugins>
             <plugin>
@@ -33,12 +31,6 @@
         <dependency>
             <groupId>org.opennms.plugins.velocloud</groupId>
             <artifactId>velocloud-client-api</artifactId>
-            <version>0.1.0-SNAPSHOT</version>
-            <scope>compile</scope>
-        </dependency>
-        <dependency>
-            <groupId>org.opennms.plugins.velocloud</groupId>
-            <artifactId>velocloud-client-v1</artifactId>
             <version>0.1.0-SNAPSHOT</version>
             <scope>compile</scope>
         </dependency>
