+++ client/cache/src/main/java/org/opennms/velocloud/client/cache/ApiCallExecution.java
@@ -26,16 +26,17 @@
  *     http://www.opennms.com/
  *******************************************************************************/
 
-package org.opennms.velocloud.cache;
+package org.opennms.velocloud.client.cache;
 
 import org.opennms.velocloud.client.api.VelocloudApiException;
 
 /**
- * Method specification that have one Parameter, returns a result and can throw a VelocloudApiException
+ * Functional interface that represents first part of a method call which can be cached (or not) independent of whole
+ * method. That allows to reuse this part in different methods.
  * @param <I> type of (I)initial parameter
- * @param <R> type of the final (R)result
+ * @param <C> Typ of the result of API call (that can be (C)cacheable)
  */
 @FunctionalInterface
-public interface VelocloudFunction<I, R> {
-    R apply(I initialParameter) throws VelocloudApiException;
+public interface ApiCallExecution<I, C> {
+    C doApiCall(I initialParameter) throws VelocloudApiException;
 }
