+++ client/cache/src/main/java/org/opennms/velocloud/client/cache/ApiGetSpecificationWithCache.java
@@ -26,7 +26,7 @@
  *     http://www.opennms.com/
  *******************************************************************************/
 
-package org.opennms.velocloud.cache;
+package org.opennms.velocloud.client.cache;
 
 import java.util.function.Supplier;
