+++ client/cache/src/main/java/org/opennms/velocloud/client/cache/FunctionSpecification.java
@@ -26,7 +26,7 @@
  *     http://www.opennms.com/
  *******************************************************************************/
 
-package org.opennms.velocloud.cache;
+package org.opennms.velocloud.client.cache;
 
 import org.opennms.velocloud.client.api.VelocloudApiException;
