+++ client/cache/src/main/java/org/opennms/velocloud/client/cache/FunctionSpecificationWithCache.java
@@ -26,16 +26,30 @@
  *     http://www.opennms.com/
  *******************************************************************************/
 
-package org.opennms.velocloud.cache;
+package org.opennms.velocloud.client.cache;
+
+import java.util.function.Function;
 
 /**
- * Represents a call of a function in object of type A with parameter P. This function can throw an Exception of type E
+ * FunctionSpecificationWithCache adds a final result adapter to ApiCallSpecification what allows to use same cache for same
+ * API calls, but calculate different results.
+ *
+ * @param <I> type of (I)initial parameter
  * @param <A> (A)API Class to call its method
  * @param <P> type of (P)parameter for API method: A.method(P param)
- * @param <C> Typ of the (C)cacheable result of API call
+ * @param <C> typ of the (C)cacheable result of API call
+ * @param <R> type of the final (R)result
  * @param <E> type of the (E)exception that can be thrown in A.apiCall(P)
  */
-@FunctionalInterface
-public interface ApiCall<A, P, C, E extends Exception> {
-    C doCall(final A api, final P parameter) throws E;
+public class FunctionSpecificationWithCache<I, A, P, C, R, E extends Exception>
+        extends FunctionSpecification<I, C, R> {
+
+    public FunctionSpecificationWithCache(
+            final Function<I, ParamsForApiCall<A, P>> prepare,
+            final ApiCall<A, P, C, E> apiCall,
+            String desc,
+            final ResultAdapter<C, R> adapter
+    ) {
+        super(new ApiCallSpecificationWithCache<>(prepare, apiCall, desc), adapter);
+    }
 }
