+++ client/cache/src/main/java/org/opennms/velocloud/client/cache/VelocloudFunction.java
@@ -26,33 +26,16 @@
  *     http://www.opennms.com/
  *******************************************************************************/
 
-package org.opennms.velocloud.cache.clieant;
+package org.opennms.velocloud.client.cache;
 
-import java.util.List;
-
-import org.opennms.velocloud.cache.VelocloudSupplier;
-import org.opennms.velocloud.client.api.VelocloudApiCustomerClient;
 import org.opennms.velocloud.client.api.VelocloudApiException;
-import org.opennms.velocloud.client.api.model.Edge;
-import org.opennms.velocloud.client.api.model.User;
 
-public abstract class VelocloudApiCustomerClientUsingSpecs implements VelocloudApiCustomerClient {
-
-    private final VelocloudSupplier<List<Edge>> getEdgesSpecification;
-    private final VelocloudSupplier<List<User>> getUsersSpecification;
-
-    public VelocloudApiCustomerClientUsingSpecs(VelocloudSupplier<List<Edge>> getEdgesSpecification, VelocloudSupplier<List<User>> getUsersSpecification) {
-        this.getEdgesSpecification = getEdgesSpecification;
-        this.getUsersSpecification = getUsersSpecification;
-    }
-
-    @Override
-    public List<Edge> getEdges() throws VelocloudApiException {
-        return getEdgesSpecification.get();
-    }
-
-    @Override
-    public List<User> getUsers() throws VelocloudApiException {
-        return getUsersSpecification.get();
-    }
+/**
+ * Method specification that have one Parameter, returns a result and can throw a VelocloudApiException
+ * @param <I> type of (I)initial parameter
+ * @param <R> type of the final (R)result
+ */
+@FunctionalInterface
+public interface VelocloudFunction<I, R> {
+    R apply(I initialParameter) throws VelocloudApiException;
 }
