+++ client/cache/src/main/java/org/opennms/velocloud/client/cache/base/VelocloudApiPartnerClientUsingSpecs.java
@@ -25,12 +25,12 @@
  *     http://www.opennms.org/
  *     http://www.opennms.com/
  *******************************************************************************/
-package org.opennms.velocloud.cache.clieant;
+package org.opennms.velocloud.client.cache.base;
 
 import java.util.List;
 import java.util.function.Function;
 
-import org.opennms.velocloud.cache.VelocloudSupplier;
+import org.opennms.velocloud.client.cache.VelocloudSupplier;
 import org.opennms.velocloud.client.api.VelocloudApiCustomerClient;
 import org.opennms.velocloud.client.api.VelocloudApiException;
 import org.opennms.velocloud.client.api.VelocloudApiPartnerClient;
